requirements.txt
dominate==2.6.0
tqdm==4.63.1

# Base
imageio==2.16.2
numpy==1.22.3
opencv-python==4.5.5.62
Pillow==9.1.0
PyYAML==6.0
requests==2.27.1
scipy>=1.4.1
torch==1.11.0
torchvision==0.11.3
tqdm>=4.41.0

# Logging
tensorboard>=2.4.1
# wandb

# Plotting
pandas>=1.1.4
matplotlib==3.5.1
seaborn>=0.11.0

# Storage
boto3==1.20.26
python-dotenv==0.19.2


